Fix MSVC C4244 warning in format specifiers

+++ include/etl/basic_format_spec.h
@@ -137,13 +137,13 @@ namespace etl
   //***************************************************************************
   static ETL_CONSTEXPR private_basic_format_spec::base_spec setbase(uint32_t base)
   {
-    return private_basic_format_spec::base_spec(base);
+    return private_basic_format_spec::base_spec(static_cast<uint_least8_t>(base));
   }
 
   //*********************************
   static ETL_CONSTEXPR private_basic_format_spec::width_spec setw(uint32_t width)
   {
-    return private_basic_format_spec::width_spec(width);
+    return private_basic_format_spec::width_spec(static_cast<uint_least8_t>(width));
   }
 
   //*********************************
@@ -156,20 +156,20 @@ namespace etl
   //*********************************
   static ETL_CONSTEXPR private_basic_format_spec::precision_spec setprecision(uint32_t precision)
   {
-    return private_basic_format_spec::precision_spec(precision);
+    return private_basic_format_spec::precision_spec(static_cast<uint_least8_t>(precision));
   }
 
   //*********************************
-  static ETL_CONSTANT private_basic_format_spec::base_spec bin(2U);
+  static ETL_CONSTANT private_basic_format_spec::base_spec bin(static_cast<uint_least8_t>(2U));
 
   //*********************************
-  static ETL_CONSTANT private_basic_format_spec::base_spec oct(8U);
+  static ETL_CONSTANT private_basic_format_spec::base_spec oct(static_cast<uint_least8_t>(8U));
 
   //*********************************
-  static ETL_CONSTANT private_basic_format_spec::base_spec dec(10U);
+  static ETL_CONSTANT private_basic_format_spec::base_spec dec(static_cast<uint_least8_t>(10U));
 
   //*********************************
-  static ETL_CONSTANT private_basic_format_spec::base_spec hex(16U);
+  static ETL_CONSTANT private_basic_format_spec::base_spec hex(static_cast<uint_least8_t>(16U));
 
   //*********************************
   static ETL_CONSTANT private_basic_format_spec::left_spec left = private_basic_format_spec::left_spec();
